button was in wrong part of component moved to the right spot
