+++ client/components/complete-itinerary/CompleteItinerary.jsx
@@ -1,10 +1,33 @@
-import React from "react";
+import React, { useState } from "react";
 import { useSelector } from "react-redux";
 import Header from "../header";
 import SingleDayItinerary from "../single-day-itinerary";
 
 const CompleteItinerary = () => {
   const itinerary = useSelector((state) => state.itinerary.itinerary);
+  const [editedItinerary, setEditedItinerary] = useState({ itinerary });
+
+  //=======> HANDLE CLICK <============
+  const handleClick = async () => {
+    console.log("state to send to back end", editedItinerary);
+    try {
+      const response = await fetch("/api/trip/update", {
+        method: "PATCH",
+        headers: {
+          "Content-Type": "application/json",
+          Authorization: `Bearer ${localStorage.getItem("userToken")}`,
+        },
+        body: JSON.stringify(editedItinerary),
+      });
+      if (response.ok) {
+        console.log("successful patch");
+      } else {
+        throw new Error("failed to retrieve data");
+      }
+    } catch (error) {
+      console.error("Error with patch request:", error);
+    }
+  };
 
   let dates;
   let dateComponents;
@@ -18,9 +41,10 @@ const CompleteItinerary = () => {
           <h2 className="date">{date}</h2>
           <div className="day-details">
             <SingleDayItinerary
-              itinerary={itinerary}
+              editedItinerary={editedItinerary}
               dateObj={itinerary[date]}
               date={date}
+              setEditedItinerary={setEditedItinerary}
             />
           </div>
         </div>
@@ -39,6 +63,7 @@ const CompleteItinerary = () => {
     <div>
       <Header />
       {itineraryItems}
+      <button onClick={handleClick}>save</button>
     </div>
   );
 };
