+++ client/components/single-day-itinerary/SingleDayItinerary.jsx
@@ -1,10 +1,14 @@
 import React, { useState } from "react";
 
-const SingleDayItinerary = ({ itinerary, dateObj, date }) => {
+const SingleDayItinerary = ({
+  setEditedItinerary,
+  editedItinerary,
+  dateObj,
+  date,
+}) => {
   const timeSlots = Object.keys(dateObj);
   const timeSlotComponents = timeSlots.map((timeOfDay) => {
     const { activity, description, address } = dateObj[timeOfDay];
-    const [editedItinerary, setEditedItinerary] = useState({ itinerary });
     //=======> HANDLE CHANGE <============
     const handleChange = (date, timeOfDay, field, e) => {
       const ItineraryCopy = JSON.parse(JSON.stringify(editedItinerary));
@@ -16,27 +20,7 @@ const SingleDayItinerary = ({ itinerary, dateObj, date }) => {
       ItineraryCopy.itinerary[date][timeOfDay][field] = e;
       setEditedItinerary(ItineraryCopy);
       console.log("edit IT--->", editedItinerary);
-    };
-    //=======> HANDLE CLICK <============
-    const handleClick = async () => {
-      console.log("state to send to back end", editedItinerary);
-      try {
-        const response = await fetch("/api/trip/update", {
-          method: "PATCH",
-          headers: {
-            "Content-Type": "application/json",
-            Authorization: `Bearer ${localStorage.getItem("userToken")}`,
-          },
-          body: JSON.stringify(editedItinerary),
-        });
-        if (response.ok) {
-          console.log("successful patch");
-        } else {
-          throw new Error("failed to retrieve data");
-        }
-      } catch (error) {
-        console.error("Error with patch request:", error);
-      }
+      return editedItinerary;
     };
     //=======> COMPONENT <============
     return (
@@ -69,7 +53,6 @@ const SingleDayItinerary = ({ itinerary, dateObj, date }) => {
           }
           id="Address"
         ></input>
-        <button onClick={handleClick}>save</button>
       </div>
     );
   });
